move parse logic into a dedicated module
